WiP: Refresh send-tasks script.
